+++ src/job_state_change.jl
@@ -130,7 +130,7 @@ function unsafe_run!(job::Job, current::DateTime=now()) :: UInt8
         return SKIP
     else  # FAIL
         # check status when fail
-        if istaskfailed2(job.task)
+        if istaskfailed(job.task)
             if job.state !== CANCELLED
                 job.state = FAILED
                 @error "A job has failed: $(job.id)" exception=job.task.result
@@ -184,7 +184,7 @@ function unsafe_cancel!(job::Job, current::DateTime=now())
     if !istaskstarted(job.task)
         job.state = CANCELLED
         return job.state
-    elseif JobSchedulers.istaskfailed2(job.task)
+    elseif istaskfailed(job.task)
         if job.state !== CANCELLED
             job.state = FAILED
             # job.task.result isa Exception, notify errors
@@ -211,53 +211,4 @@ function unsafe_cancel!(job::Job, current::DateTime=now())
     finally
         return job.state
     end
-end
-
-
-"""
-    istaskfailed2(t::Task)
-
-Extend `Base.istaskfailed` to fit Pipelines and JobSchedulers packages, which will return a `StackTraceVector` in `t.result`, while Base considered it as `:done`. The function checks the situation and modifies the real task status and other properties.
-"""
-function istaskfailed2(t::Task)
-    if Base.istaskfailed(t)
-        return true
-    end
-    @static if hasfield(Task, :_state)
-        if t._state === 0x02 # Base.task_state_failed
-            return true
-        end
-    elseif hasfield(Task, :state)
-        # TODO: this field name should be deprecated in 2.0
-        if t.state === :failed
-            return true
-        end
-    end
-    if getproperty(t, :result) isa Pipelines.StackTraceVector
-        # it is failed, but task is showing done, so we make it failed manually.
-        @static if hasfield(Task, :_state)
-            @static if VERSION >= v"1.12-"  # 1.12: t._state is atomic
-                @atomic t._state = 0x02  # Base.task_state_failed
-            else
-                t._state = 0x02  # Base.task_state_failed
-            end
-        end
-        @static if hasfield(Task, :_isexception)
-            t._isexception = true
-        end
-        @static if hasfield(Task, :state)
-            # TODO: this field name should be deprecated in 2.0
-            t.state = :failed
-        end
-        @static if hasfield(Task, :exception)
-            # TODO: this field name should be deprecated in 2.0
-            t.exception = t.result
-        end
-        @static if hasfield(Task, :backtrace)
-            # TODO: this field name should be deprecated in 2.0
-            t.backtrace = t.result[end][2]
-        end
-        return true
-    end
-    return false
 end