+++ docs/src/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+v0.12.0-dev
+
+- Compat: ScopedStreams v0.3.6: Allow job-dependent stdout and stderr.
+- Compat: Pipelines v0.12: `StackTraceVector` is no longer available, so `istaskfailed2` is not needed.
+- Breaking: method deleted: `istaskfailed2`.
+
 v0.11.12
 
 - Optim: `@submit`: check `expr isa Symbol` during compilation.